Fix server:delete when installer was helm

+++ src/tasks/installers/operator.ts
@@ -355,11 +355,15 @@ export class OperatorTasks {
     {
       title: `Delete CRD ${this.cheClusterCrd}`,
       task: async (_ctx: any, task: any) => {
+        const crdExists = await kh.crdExist(this.cheClusterCrd)
         const checlusters = await kh.getAllCheCluster()
         if (checlusters.length > 0) {
           task.title = await `${task.title}...Skipped: another Eclipse Che deployment found.`
         } else {
-          await kh.deleteCrd(this.cheClusterCrd)
+          // Check if CRD exist. When installer is helm the CRD are not created
+          if (crdExists) {
+            await kh.deleteCrd(this.cheClusterCrd)
+          }
           task.title = await `${task.title}...OK`
         }
       }
